Add Support for Metrics
It is now possible to view the metrics for some Kubernetes resources,
like Pods, Nodes and Deployments, directly within the details view of
the Kubernetes resources. To use this feature, the Kubernetes
datasource must be configured with the UID of a Prometheus datasource
and the job label for the kubelet, kube-state-metrics, and node-exporter
metrics.