Add Label to Kubernetes Resources Queries (#38)
When we display a list of Kubernetes resources in the variables
selection, we now include their `kind` as labels. This enhancement
should improve the UI / UX becaue we do not display the raw resource ids
anymore. The ids are still used as value, to identify the correct
resource based on user selection.