Format Metrics in Resources View (#43)
The `PodMetrics` and `NodeMetrics` are now better formatted in the
resources view. Instead of directly using the result from the Kubernetes
API, we are parsing the returned values and formatting them into the
same format as `kubectl top` does.

This means we are using millicores (`m`) for CPU and mebibytes (`Mi`) for
memory.