Replace Explore Links with Drawer (#52)
Instead of linking to the explore view, we now show the resources,
linked in the details, within a drawer. This has the advantage that the
user does not lose context of the current view and that we can render
the actions for the opened resource directly in the drawer.

This commit also fixes the query for events and Helm releases.