Use Frontend for Redirect in `kubectl-grafana`
Instead of directly calling the API of the Kubernetes datasource to get
a Kubeconfig / credentials, we are now opening the Grafana frontend,
which calls the API to get the data and then redirects to the
`kubectl-grafana` command.

This is done, because when we directly call the API and the user is not
authenticated, Grafana will return an error. When we open the Grafana
frontend, the user is redirect to the authentication page first.

This commit also renames the `token` command to `credentials`.